pyproject.toml
[tool.poetry]
name = "flux_server"
version = "0.1.0"
description = ""
authors = ["Sarath S Menon <sarathmenon.downey@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
torch="2.4.1"
torchvision="0.19.1"
safetensors="0.4.4"
diffusers="0.30.0"
transformers="4.44.0"
lycoris-lora="1.8.3"
flatten-json="0.1.14"
pyyaml="6.0.1"
oyaml="1.0"
tensorboard="2.17.0"
kornia="0.7.3"
einops="0.8.0"
accelerate="0.33.0"
toml="0.10.2"
albumentations="1.4.3"
opencv-python="4.10.0.84"
pillow="10.4.0"
pydantic="1.10.17"
omegaconf="2.3.0"
open-clip-torch="2.26.1"
timm="1.0.8"
prodigyopt="1.0"
controlnet-aux="0.0.7"
scikit-image="0.24.0"
huggingface-hub="0.24.5"
hf-transfer="0.1.8"
lpips="0.1.4"
optimum-quanto="0.2.4"
sentencepiece="0.2.0"
peft="0.13.0"
lion-pytorch="0.2.2"
openai="1.58.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
